requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
altair>=4.2.0
psycopg2-binary>=2.9.0
bcrypt>=4.0.0
google-auth>=2.0.0
google-auth-oauthlib>=0.5.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0